+++ mindee/v2/commands/__init__.py
@@ -1,16 +1,22 @@
+from mindee.v2.commands.base_inference_command import BaseInferenceCommand
+from mindee.v2.commands.classification_command import ClassificationCommand
 from mindee.v2.commands.cli_parser import MindeeArgumentParser, MindeeParser
-from mindee.v2.commands.inference_command import (
-    InferenceCommand,
-    InferenceCommandOptions,
-)
+from mindee.v2.commands.crop_command import CropCommand
+from mindee.v2.commands.extraction_command import ExtractionCommand
+from mindee.v2.commands.ocr_command import OcrCommand
 from mindee.v2.commands.output_type import OutputType
 from mindee.v2.commands.search_models_command import SearchModelsCommand
+from mindee.v2.commands.split_command import SplitCommand
 
 __all__ = [
-    "InferenceCommand",
-    "InferenceCommandOptions",
+    "BaseInferenceCommand",
+    "ClassificationCommand",
+    "CropCommand",
+    "ExtractionCommand",
     "MindeeArgumentParser",
     "MindeeParser",
+    "OcrCommand",
     "OutputType",
     "SearchModelsCommand",
+    "SplitCommand",
 ]